requirements.txt
streamlit
langchain
langchain-groq
langgraph
requests
python-dotenv
pandas
mcp
aiohttp
fastapi
fpdf
groq
uvicorn
reportlab
xlsxwriter